error fixed, changed the createMockRouter to .tsx
